% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_warnings.R
\name{catch_warnings}
\alias{catch_warnings}
\title{catch warnings alongside results without returning warning}
\source{
\code{demo(error.catching)}
}
\usage{
catch_warnings(expr)
}
\arguments{
\item{expr}{an \R expression to evaluate}
}
\value{
a list where the first entry is the result of \code{expr} and the
second provides information about warnings
}
\description{
helper function
}
\examples{
log(3)
catch_warnings(log(3))

# produces warning
# log(-3)
# catch it
catch_warnings(log(-3))

# produces error
# log("x")
# catch it
catch_warnings(log("x"))
}
