% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_steepness_from_matrix.R
\name{elo_steepness_from_matrix}
\alias{elo_steepness_from_matrix}
\title{steepness based on Bayesian Elo-rating}
\usage{
elo_steepness_from_matrix(
  mat,
  algo = c("fixed_sd", "original"),
  n_rand = NULL,
  silent = FALSE,
  static = TRUE,
  ...
)
}
\arguments{
\item{mat}{square interaction matrix}

\item{algo}{character, either \code{"fixed_sd"} or \code{"original"}.
This determines which algorithm is used. Default is
\code{"fixed_sd"}, which is a slight modification from
Goffe et al's original code.}

\item{n_rand}{numeric, number of randomized sequences. Default is
\code{NULL}, which uses a rule of thumb to determine the
number (see below for more details).}

\item{silent}{logical, suppress warnings}

\item{static}{logical, treat the data as static (i.e. think of data in
matrix form). Default is \code{TRUE}. \code{FALSE} is not
yet supported.}

\item{...}{additional arguments for \code{sampling()}}
}
\value{
a list
}
\description{
steepness based on Bayesian Elo-rating
}
\details{
The number of randomizations is set in the following way, unless
a specific number is provided. If there are more than 500
observed interactions, \code{n_rand = 5}. If there are less than
100 interactions, \code{n_rand = 50}. In the remaining cases,
\code{n_rand = 20}.
}
\examples{
data(dommats, package = "EloRating")
\dontrun{
res <- elo_steepness_from_matrix(dommats$elephants, n_rand = 3, cores = 4,
                                 iter = 3000, warmup = 2000, refresh = 0)
plot_steepness(res)

res <- elo_steepness_from_matrix(dommats$elephants, n_rand = 3, cores = 4,
                                 algo = "original",
                                 iter = 3000, warmup = 2000, refresh = 0)
plot_steepness(res)
}
}
