% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{plot (rather than print) a matrix}
\usage{
plot_matrix(mat, greyout = NULL, prunkcol = NULL)
}
\arguments{
\item{mat}{square matrix}

\item{greyout}{numeric, the values to be grayed out}

\item{prunkcol}{color value, which if set to some color will highlight
unknown relationships with rectangles of that color.}
}
\value{
a plot
}
\description{
a helper function
}
