% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_for_rstan.R
\name{prep_data_for_rstan}
\alias{prep_data_for_rstan}
\title{prepare data for stan call}
\usage{
prep_data_for_rstan(mat, n_rand = 1, silent = FALSE, for_elo_model = TRUE)
}
\arguments{
\item{mat}{square interaction matrix}

\item{n_rand}{numeric, number of randomizations}

\item{silent}{logical, omit printing messages regarding non-fatal data
issues. Default is \code{FALSE}, i.e. do print messages.}

\item{for_elo_model}{logical, output ready for Elo steepness (default,
\code{TRUE}). If \code{FALSE}, prep for David's score steepness.}
}
\value{
a list that is formatted so that it can be handed over to the
respective Stan models
}
\description{
prepare data for stan call
}
