% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatability_steepness.R
\name{repeatability_steepness}
\alias{repeatability_steepness}
\title{steepness via repeatability (cf aniDom package)}
\usage{
repeatability_steepness(mat, n_rand = 1000)
}
\arguments{
\item{mat}{square interaction matrix}

\item{n_rand}{numeric, number of randomized sequences
(default is \code{1000})}
}
\value{
a steepness value
}
\description{
steepness via repeatability (cf aniDom package)
}
\examples{
data(bonobos, package = "EloRating")
repeatability_steepness(bonobos, n_rand = 20)
}
\references{
Sanchez-Tojar et al 2018
}
