% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler_diagnostics.R
\name{sampler_diagnostics}
\alias{sampler_diagnostics}
\title{catch Stan sampling issues without throwing a warning}
\usage{
sampler_diagnostics(object)
}
\arguments{
\item{object}{\code{\link[rstan]{stanfit}} object}
}
\value{
a list regarding any sampling issues encountered during fitting
}
\description{
catch Stan sampling issues without throwing a warning
}
