% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upward_steepness.R
\name{upward_steepness}
\alias{upward_steepness}
\title{proportion of interactions against the rank order}
\usage{
upward_steepness(mat)
}
\arguments{
\item{mat}{square interaction matrix}
}
\value{
numeric value of upward steepness
}
\description{
proportion of interactions against the rank order
}
\examples{
data(bonobos, package ="EloRating")
upward_steepness(bonobos)
}
