% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davids_steepness.R
\name{davids_steepness}
\alias{davids_steepness}
\title{David's scores and steepness with Bayesian flavor}
\usage{
davids_steepness(mat, silent = FALSE, ...)
}
\arguments{
\item{mat}{square interaction matrix}

\item{silent}{logical, suppress warnings (default is \code{FALSE})}

\item{...}{additional arguments for \code{\link[rstan]{sampling}()}}
}
\value{
a list or stan fit object
}
\description{
David's scores and steepness with Bayesian flavor
}
\examples{
\dontrun{
data(dommats, package = "EloRating")
res <- davids_steepness(dommats$elephants, refresh = 0)
plot_steepness(res)
}
}
