% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{plot (rather than print) a matrix}
\usage{
plot_matrix(mat, greyout = NULL, prunkcol = NULL, label_col = "black")
}
\arguments{
\item{mat}{square matrix}

\item{greyout}{numeric, the values to be grayed out}

\item{prunkcol}{color value, which if set to some color will highlight
unknown relationships with rectangles of that color.}

\item{label_col}{color values for column and row labels}
}
\value{
a plot and an invisible list with coordinates and
content of the matrix to be plotted
}
\description{
a helper function
}
