% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.elo_steepness}
\alias{summary.elo_steepness}
\alias{summary.david_steepness}
\title{summary}
\usage{
\method{summary}{elo_steepness}(object, ...)

\method{summary}{david_steepness}(object, ...)
}
\arguments{
\item{object}{result from \code{\link{elo_steepness_from_matrix}},
\code{\link{elo_steepness_from_sequence}} or
\code{\link{davids_steepness}}}

\item{...}{further arguments passed to or from other methods (ignored)}
}
\value{
Nothing returned. Called for side effects of textual output
to console.
}
\description{
summary
}
